Flip bits

+++ 0x14-bit_manipulation/5-flip_bits.c
@@ -1,7 +1,7 @@
 #include "main.h"
 
 /**
- * flip_bits - Checks and returns the number of 
+ * flip_bits - Checks and returns the number of
  * bits you would need to flip to get from one number to another.
  * @n: First number
  * @m: Second number
@@ -10,7 +10,7 @@
  */
 unsigned int flip_bits(unsigned long int n, unsigned long int m)
 {
-	unsigned int count = 0;
+	unsigned long int count = 0;
 	unsigned int xor = n ^ m;
 
 	while (xor > 0)
